✨ (flatpak-sync) Fail add/remove if app id is commented out
Handling commented out apps would perhaps be interesting to add in the
future. But at this point I think the the people using this script
either want to go fully declarative and not use the add/remove
commands at all or do not want to edit the apps list file themselves
and therefore will not have and comments in them.

Refs: #1